+++ src/main/java/net/grid/vampiresdelight/common/registry/VDBlockEntityTypes.java
@@ -1,7 +1,6 @@
 package net.grid.vampiresdelight.common.registry;
 
 import net.grid.vampiresdelight.VampiresDelight;
-import net.grid.vampiresdelight.common.block.entity.DarkStoneStoveBlockEntity;
 import net.grid.vampiresdelight.common.block.entity.WineShelfBlockEntity;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -17,8 +16,6 @@ import java.util.function.Supplier;
 public class VDBlockEntityTypes {
     public static final DeferredRegister<BlockEntityType<?>> BLOCK_ENTITIES = DeferredRegister.create(Registries.BLOCK_ENTITY_TYPE, VampiresDelight.MODID);
 
-    public static final Supplier<BlockEntityType<DarkStoneStoveBlockEntity>> DARK_STONE_STOVE = BLOCK_ENTITIES.register("dark_stone_stove",
-            () -> BlockEntityType.Builder.of(DarkStoneStoveBlockEntity::new, VDBlocks.DARK_STONE_STOVE.get()).build(null));
     public static final Supplier<BlockEntityType<WineShelfBlockEntity>> WINE_SHELF = BLOCK_ENTITIES.register("wine_shelf",
             () -> BlockEntityType.Builder.of(WineShelfBlockEntity::new,
                             VDBlocks.OAK_WINE_SHELF.get(),
@@ -46,5 +43,7 @@ public class VDBlockEntityTypes {
                 VDBlocks.JACARANDA_CABINET.get(),
                 VDBlocks.MAGIC_CABINET.get()
         );
+        event.modify(ModBlockEntityTypes.STOVE.get(),
+                VDBlocks.DARK_STONE_STOVE.get());
     }
 }